signatures: Move functions tests to a separate file
